Update: Rename all module to snake case.
