+++ src/smart_qq_bot/msg_handler.py
@@ -3,9 +3,9 @@
 # Code by Yinzo:        https://github.com/Yinzo
 # Origin repository:    https://github.com/Yinzo/SmartQQBot
 
-from Group import *
-from Pm import *
-from Sess import *
+from group import *
+from pm import *
+from sess import *
 import threading
 
 logging.basicConfig(
